Re-organise and polish DVL documentation
We heavily re-organise to try to avoid duplication, and to try to
enable info to be found as easily as possible. In particular, we extract certain
pieces of info into stand-alone pages, and move some pages up a level.

We essentially rewrite some pages, and heavily edit the langauge of
almost all of them, to try to make things clearer and more easily
digestible.