Fix locator page structure and rewrite UGPS API page
Previously, the locator links in the G2 manual menu ended up being
displayed as belonging to the Explorer Kit manual. We fix this, and in
addition rewrite the page on the UGPS API to improve the language,
remove usage details (as this covered at github), and provide a
summary of what is covered in the linked-to usage examples.